+++ src/main/java/com/tlabs/eve/api/character/CharacterCalendarRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.character;
 
 public final class CharacterCalendarRequest extends CharacterRequest<CharacterCalendarResponse> {
 
-    public static final int MASK = 1048576;
+    public static final long MASK = 1048576;
 
     public CharacterCalendarRequest(String characterID) {
         super(CharacterCalendarResponse.class, "/char/UpcomingCalendarEvents.xml.aspx", MASK, characterID);
