+++ src/main/java/com/tlabs/eve/api/character/CharacterTrainingRequest.java
@@ -6,7 +6,7 @@ package com.tlabs.eve.api.character;
 
 
 public final class CharacterTrainingRequest extends CharacterRequest<CharacterTrainingResponse> {
-    public static final int MASK = 131072;
+    public static final long MASK = 131072;
 
     public CharacterTrainingRequest(String charID) {
         super(CharacterTrainingResponse.class, "/char/SkillInTraining.xml.aspx", MASK, charID);
