+++ src/main/java/com/tlabs/eve/api/corporation/CorporationContractItemsRequest.java
@@ -5,7 +5,7 @@ package com.tlabs.eve.api.corporation;
 import com.tlabs.eve.api.ContractItemsResponse;
 
 public final class CorporationContractItemsRequest extends CorporationRequest<ContractItemsResponse> {
-    public static final int MASK = 8388608;
+    public static final long MASK = 8388608;
 
     private final long contractID;
