+++ src/main/java/com/tlabs/eve/api/corporation/StarbaseListRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.corporation;
 
 
 public final class StarbaseListRequest extends CorporationRequest<StarbaseListResponse> {
-    public static final int MASK = 524288;
+    public static final long MASK = 524288;
 
     public StarbaseListRequest(String corporationID) {
         super(StarbaseListResponse.class, "/corp/StarbaseList.xml.aspx", MASK, corporationID);
