+++ src/main/java/com/tlabs/eve/api/mail/MailingListsRequest.java
@@ -6,7 +6,7 @@ import com.tlabs.eve.api.character.CharacterRequest;
 
 ///char/mailinglists.xml.aspx
 public final class MailingListsRequest extends CharacterRequest<MailingListsResponse> {
-    public static final int MASK = 1024;
+    public static final long MASK = 1024;
 
     public MailingListsRequest(String characterID) {
         super(MailingListsResponse.class, "/char/mailinglists.xml.aspx", MASK, characterID);
