+++ src/main/java/com/tlabs/eve/api/mail/NotificationsRequest.java
@@ -6,7 +6,7 @@ import com.tlabs.eve.api.character.CharacterRequest;
 
 ///char/MailMessages.xml.aspx
 public final class NotificationsRequest extends CharacterRequest<NotificationsResponse> {
-    public static final int MASK = 16384;
+    public static final long MASK = 16384;
 
     public NotificationsRequest(String characterID) {
         super(NotificationsResponse.class, "/char/Notifications.xml.aspx", MASK, characterID);
